feat: complete user menu options
Now:
-prompt question before doing major processes
-allow user to abort mid sessions
-user can now use atm atm operations
-log out option
